(refactor) Reorganize type definitions for better separation of concerns
Summary
* Created dedicated osc-listener.d.ts for internal type definitions
* Cleaned up osc.d.ts to only contain external module declarations
* Fixed potential issues with null/undefined values

Details
* Created new osc-listener.d.ts file that contains all internal type definitions
* Removed internal type definitions from osc.d.ts
* Eliminated redundant custom UDPPort interface
* Updated imports across the codebase to reference correct type definition files
* Added null checking for recordFileName to prevent runtime errors
* Updated re-exports to explicitly include the UDPPort from 'osc' module
